New: Terminer parser declaration + tests

+++ src/elements/formes/Carre.h
@@ -4,13 +4,13 @@
 
 class Carre: public Forme {
 	public:
-		Carre(int x1, int y1, int taille) noexcept: Forme(Type::Carre), _x1(x1), _y1(y1), _taille(taille) {}
-		Carre(Proprietes prop, int x1, int y1, int taille) noexcept: Forme(prop, Type::Carre), _x1(x1), _y1(y1), _taille(taille) {}
+		Carre(double x1, double y1, double taille) noexcept: Forme(Type::Carre), _x1(x1), _y1(y1), _taille(taille) {}
+		Carre(Proprietes prop, double x1, double y1, double taille) noexcept: Forme(prop, Type::Carre), _x1(x1), _y1(y1), _taille(taille) {}
 
 		std::string to_svg() const override { return ""; }
 
 	private:
-		int _x1;
-		int _y1;
-		int _taille;
+		double _x1;
+		double _y1;
+		double _taille;
 };
