+++ src/elements/formes/Chemin.cpp
@@ -1,5 +1,5 @@
 #include "Chemin.h"
 
-void Chemin::ajoutePoint(int x, int y) {
+void Chemin::ajoutePoint(double x, double y) {
     _points.push_back(Point {x, y});
 }
