+++ src/elements/formes/Chemin.h
@@ -6,17 +6,17 @@
 
 class Chemin: public Forme {
 	struct Point {
-		int x;
-		int y;
+		double x;
+		double y;
 	};
 
 	public:
-	Chemin(int x1, int y1) noexcept: Forme(Type::Chemin), _points(1, Point{ x1, y1 }) {}
-	Chemin(Proprietes prop, int x1, int y1) noexcept: Forme(prop, Type::Chemin), _points(1, Point{ x1, y1 }) {}
+	Chemin(double x1, double y1) noexcept: Forme(Type::Chemin), _points(1, Point{ x1, y1 }) {}
+	Chemin(Proprietes prop, double x1, double y1) noexcept: Forme(prop, Type::Chemin), _points(1, Point{ x1, y1 }) {}
 
 	std::string to_svg() const override { return ""; }
 
-	void ajoutePoint(int x, int y);
+	void ajoutePoint(double x, double y);
 
 	private:
 	std::vector<Point> _points;
