+++ src/elements/formes/Rectangle.h
@@ -4,11 +4,11 @@
 
 class Rectangle: public Forme {
 public:
-    Rectangle(int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4) noexcept: Forme(Type::Rectangle), _coordonnees { x1, y1, x2, y2, x3, y3, x4, y4 } {}
-    Rectangle(Proprietes prop, int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4) noexcept: Forme(prop, Type::Rectangle), _coordonnees { x1, y1, x2, y2, x3, y3, x4, y4 } {}
+    Rectangle(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4) noexcept: Forme(Type::Rectangle), _coordonnees { x1, y1, x2, y2, x3, y3, x4, y4 } {}
+    Rectangle(Proprietes prop, double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4) noexcept: Forme(prop, Type::Rectangle), _coordonnees { x1, y1, x2, y2, x3, y3, x4, y4 } {}
 
     std::string to_svg() const override;
 
 private:
-    int _coordonnees[8];
+    double _coordonnees[8];
 };
